add check-username / check-email

+++ server.js
@@ -18,26 +18,47 @@ app.get('/', (req, res) => {
 });
 
 app.post('/register', (req, res) => {
-    const clientIp = reqip.getClientIp(req);
+
+    // search for existing username or email
+    db.accounts.findOne({ where: {
+        [db.Sequelize.Op.or]: [
+          { username: req.body.username },
+          { email: req.body.email }
+        ]
+    }})
+    .then(result => {
+        // if one of them is found
+        if (result !== null) {
+            // if the match cause is the username
+            if (result.username == req.body.username) {
+                console.log(`username ${result.username} is already used`);
+            } else {
+                console.log(`email ${result.email} is already used`);
+            }
+        } else {
+            // account registration
+            console.log('Ok !');
+
+            // force: true will drop the table if it already exists
+            db.sequelize.sync({force: false})
+            .then(() => db.accounts.create({
+                username: req.body.username,
+                email: req.body.email,
+                password: sha512(req.body.password.toLowerCase()).toString('hex'),
+                authority: 0,
+                emailRegistration: req.body.email,
+                registrationIp: reqip.getClientIp(req),
+            }))
+            .then(e => {
+                console.log(e.toJSON());
+            });
+
+            res.redirect('/success');
+        }
+    });
+
     
-    db.sequelize.sync({ force: false })
-        .then(() => db.accounts.create({
-            username: req.body.username,
-            email: req.body.email,
-            password: sha512(req.body.password).toString('hex'),
-            registrationIP: clientIp,
-            emailRegistration: req.body.email,
-            authority: 0
-        }))
-        .then(e => {
-            console.log(e.toJSON());
-        });
-
-   
-
-    res.redirect('/success');
 });
-
 app.get('/success', (req, res) => {
     res.sendFile(__dirname + "/success.html");
 })
